Fixed all issues regarding global instances

+++ src/server/db/interfaces/databaseUserInteractions.ts
@@ -1,4 +1,4 @@
-import { db } from "@/server/core/singleton";
+import { getDb } from "@/server/core/singleton";
 import { NewUser, users } from "@/shared/schemas";
 import { eq } from "drizzle-orm";
 import { User } from "@/shared/schemas";
@@ -9,8 +9,11 @@ export const DatabaseUserInteractions = {
 
   /** Used to create a user */
   async createUser(config: NewUser): Promise<User> {
-     const user = await db.insert(users).values(config).returning();
-     return user[0];
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
+
+    const user = await db.insert(users).values(config).returning();
+    return user[0];
   },
 
   // END: CREATE
@@ -19,22 +22,28 @@ export const DatabaseUserInteractions = {
 
   /** Used to retrieve a single user via their id */
   async getUserById(id: string): Promise<User | null> {
-    const result = await db.select().from(users).where(eq(users.id, id));
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
 
+    const result = await db.select().from(users).where(eq(users.id, id));
     return result[0] || null;
   },
 
   /** Used to retrieve a user via their email */
   async getUserByEmail(email: string): Promise<User | null> {
-    const result = await db.select().from(users).where(eq(users.email, email));
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
 
+    const result = await db.select().from(users).where(eq(users.email, email));
     return result[0] || null;
   },
 
   /** Used to retrieve a user via their username */
   getUserByUsername(username: string): Promise<User | null> {
-    const result = db.select().from(users).where(eq(users.username, username));
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
 
+    const result = db.select().from(users).where(eq(users.username, username));
     return result[0] || null;
   },
 
@@ -44,6 +53,9 @@ export const DatabaseUserInteractions = {
 
   /** Used to update a user's email via their id */
   async updateUserEmail(email: string, id: string): Promise<User | null> {
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
+
     const result = await db
       .update(users)
       .set({ email })
@@ -55,6 +67,9 @@ export const DatabaseUserInteractions = {
 
   /** Used to update a user's name via their id */
   async updateUserName(name: string, id: string): Promise<User | null> {
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
+
     const result = await db
       .update(users)
       .set({ name })
@@ -66,6 +81,9 @@ export const DatabaseUserInteractions = {
 
   /** Used to update a user's username via their id */
   async updateUserUsername(username: string, id: string): Promise<User | null> {
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
+
     const result = await db
       .update(users)
       .set({ username })
@@ -77,6 +95,9 @@ export const DatabaseUserInteractions = {
 
   /** Used to update a user's password via their id */
   async updateUserPassword(password: string, id: string): Promise<User | null> {
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
+
     const result = await db
       .update(users)
       .set({ password })
@@ -88,6 +109,9 @@ export const DatabaseUserInteractions = {
 
   /** Used to update a user's image via their id */
   async updateUserImage(image: string, id: string): Promise<User | null> {
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
+
     const result = await db
       .update(users)
       .set({ image })
@@ -99,6 +123,9 @@ export const DatabaseUserInteractions = {
 
   /** Used to update a user's roles via their id */
   async updateUserRole(roles: string[], id: string): Promise<User | null> {
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
+
     const result = await db
       .update(users)
       .set({ roles })
@@ -114,6 +141,9 @@ export const DatabaseUserInteractions = {
 
   /** Used to delete a user via their username*/
   async deleteUser(id: string): Promise<User | null> {
+    const db = getDb();
+    if (!db) throw new Error("Database not initialized");
+
     const result = await db.delete(users).where(eq(users.id, id)).returning();
 
     return result[0] || null;
